When registering rule on startup, take a rule int id instead of a rule when removing active rules on removed rules

+++ sonar-server/src/main/java/org/sonar/server/startup/RegisterRules.java
@@ -271,7 +271,7 @@ public final class RegisterRules {
     if (!rule.getStatus().equals(Rule.STATUS_REMOVED)) {
       LOG.info("Removing rule " + rule.ruleKey());
       profilesManager.removeActivatedRules(rule.getId());
-      rule = session.reattach(Rule.class, rule.getId());
+      rule = session.reattach(Rule.class, rule.getId().intValue());
       rule.setStatus(Rule.STATUS_REMOVED);
       rule.setUpdatedAt(new Date());
       session.save(rule);
